backend/requirements.txt
# Web框架
fastapi==0.115.4
uvicorn[standard]==0.30.6
python-multipart==0.0.9

# 配置和环境
python-dotenv==1.0.1

# 数据库
mysql-connector-python==9.0.0

# PDF处理核心依赖
PyMuPDF==1.23.14
pdfplumber==0.10.3
camelot-py[cv]==0.10.1

# 图像处理
opencv-python==4.8.1.78
Pillow==10.4.0

# OCR相关
pytesseract==0.3.10
layoutparser==0.3.4

# 计算机视觉（可选，安装较慢）
# detectron2==0.6  # 注释掉，需要特殊安装

# 数据处理
PyYAML==6.0.1
numpy==1.24.3
pandas==2.0.3

# HTTP请求
requests==2.31.0

# 日志和工具
loguru==0.7.2

# 数据校验与模型
pydantic>=2.0.0

# 其他工具库
jinja2==3.1.4